Move TrainingProviders to Providers module
