--- app/controllers/publish/training_providers/course_exports_controller.rb
@@ -1,27 +1,0 @@
-# frozen_string_literal: true
-
-module Publish
-  module TrainingProviders
-    class CourseExportsController < PublishController
-      def index
-        authorize(provider, :can_list_training_providers?)
-
-        respond_to do |format|
-          format.csv do
-            send_data(data_export.data, filename: data_export.filename, disposition: :attachment)
-          end
-        end
-      end
-
-      private
-
-      def courses
-        @courses ||= provider.current_accredited_courses
-      end
-
-      def data_export
-        @data_export ||= Exports::AccreditedCourseList.new(courses:)
-      end
-    end
-  end
-end
