+++ app/views/publish/providers/training_providers/courses/index.html.erb
