+++ app/views/publish/providers/training_providers/index.html.erb
